+++ src/platforms/esp32/components/avm_sys/sys.c
@@ -450,7 +450,8 @@ enum OpenAVMResult sys_open_avm_from_file(GlobalContext *global, const char *pat
         if (IS_NULL_PTR(part_data)) {
             return AVM_OPEN_CANNOT_OPEN;
         }
-        if (UNLIKELY(!avmpack_is_valid(part_data, size))) {
+        uint32_t avm_size;
+        if (UNLIKELY(!avmpack_compute_size(part_data, size, &avm_size))) {
             return AVM_OPEN_INVALID;
         }
 
@@ -458,7 +459,7 @@ enum OpenAVMResult sys_open_avm_from_file(GlobalContext *global, const char *pat
         if (IS_NULL_PTR(part_avm)) {
             return AVM_OPEN_FAILED_ALLOC;
         }
-        avmpack_data_init(&part_avm->base, &esp32_part_avm_pack_info);
+        avmpack_data_init(&part_avm->base, &esp32_part_avm_pack_info, avm_size);
         part_avm->base.data = part_data;
         part_avm->part_handle = part_handle;
         avmpack_data = &part_avm->base;
@@ -490,7 +491,7 @@ enum OpenAVMResult sys_open_avm_from_file(GlobalContext *global, const char *pat
             return AVM_OPEN_CANNOT_READ;
         }
 
-        if (UNLIKELY(!avmpack_is_valid(file_data, size))) {
+        if (UNLIKELY(!avmpack_is_complete(file_data, size))) {
             free(file_data);
             return AVM_OPEN_INVALID;
         }
@@ -500,7 +501,7 @@ enum OpenAVMResult sys_open_avm_from_file(GlobalContext *global, const char *pat
             free(file_data);
             return AVM_OPEN_FAILED_ALLOC;
         }
-        avmpack_data_init(&in_memory_avm->base, &in_memory_avm_pack_info);
+        avmpack_data_init(&in_memory_avm->base, &in_memory_avm_pack_info, size);
         in_memory_avm->base.data = file_data;
         avmpack_data = &in_memory_avm->base;
     }