+++ src/platforms/generic_unix/lib/sys.c
@@ -435,7 +435,8 @@ enum OpenAVMResult sys_open_avm_from_file(
     if (IS_NULL_PTR(mapped)) {
         return AVM_OPEN_CANNOT_OPEN;
     }
-    if (UNLIKELY(!avmpack_is_valid(mapped->mapped, mapped->size))) {
+    if (UNLIKELY(!avmpack_is_complete(mapped->mapped, (uint32_t) mapped->size))) {
+        mapped_file_close(mapped);
         return AVM_OPEN_INVALID;
     }
 
@@ -444,7 +445,7 @@ enum OpenAVMResult sys_open_avm_from_file(
         mapped_file_close(mapped);
         return AVM_OPEN_FAILED_ALLOC;
     }
-    avmpack_data_init(&avmpack_data->base, &mapped_file_avm_pack_info);
+    avmpack_data_init(&avmpack_data->base, &mapped_file_avm_pack_info, (uint32_t) mapped->size);
     avmpack_data->base.data = mapped->mapped;
     avmpack_data->mapped = mapped;